+++ features/pets.py
@@ -1,7 +1,7 @@
 import json
 
 import constants_parsing
-from utils.text_formatting import printSmallHeader, fmt_str
+from utils.text_utils import printSmallHeader, fmt_str
 
 
 with open('constants.json') as file:
